all working, styles need to be refactored and the tray (for palette toggle) needs to be styled
