+++ modules/@apostrophecms/admin-bar/ui/apos/components/TheAposAdminBarUser.vue
@@ -18,6 +18,7 @@
 <script>
 
 export default {
+  name: 'TheAposAdminBarUser',
   data() {
     return {
       menu: [
